refactor like functionality to use LikeArticle and LikeComment controllers, improving like handling for articles and comments

+++ resources/views/blog/add_delete_like.php
@@ -1,39 +1,64 @@
 <?php
     session_start();
 
-    require_once __DIR__ . '/../../../controllers/LikeController.php';
+    require_once __DIR__ . '/../../../controllers/LikeArticle.php';
+    require_once __DIR__ . '/../../../controllers/likeComment.php';
 
     if($_SERVER['REQUEST_METHOD'] === 'POST') {
-
-        $idUser = $_SESSION['user']['id'];
-        $idArticle = $_POST['idArticle'];
-        $currentPage = $_POST['currentPath'];
-        $redirectPage = '/resources/views/blog/blog.php';
+        
+        if(isset($_POST['likeComment'])) {
+            $idUser = $_SESSION['user']['id'];
+            $idArticle = $_POST['idArticle'];
+            $idComment = $_POST['idComment'];
+        
+            $likes = new LikeComment($idComment, $idUser);
+            $totalUserLike = $likes->totalUserLike();
+        
+            if($totalUserLike == 0) {
+                $insertLike = new LikeComment($idComment, $idUser);
     
-        $likes = new LikeController($idArticle, $idUser);
-        $totalUserLike = $likes->totalUserLike();
-    
-        if($totalUserLike == 0) {
-            $insertLike = new LikeController($idArticle, $idUser);
-
-            if($insertLike->store()) {
-                if($currentPage === $redirectPage) {
-                    header('location: blog.php');
-                } else {
+                if($insertLike->store()) {
                     header("location: detailsArticle.php?idArticle=$idArticle");
                 }
-            }
-        } else {
-            $deleteLike = new LikeController($idArticle, $idUser);
-
-            if($deleteLike->destroy()) {
-                if($currentPage === $redirectPage) {
-                    header('location: blog.php');
-                } else {
+            } else {
+                $deleteLike = new LikeComment($idComment, $idUser);
+    
+                if($deleteLike->destroy()) {
                     header("location: detailsArticle.php?idArticle=$idArticle");
                 }
-            }
-        } 
+            } 
+        } else if(isset($_POST['likeArticle'])) {
+            $idUser = $_SESSION['user']['id'];
+            $idArticle = $_POST['idArticle'];
+            $currentPage = $_POST['currentPath'];
+            $redirectPage = '/resources/views/blog/blog.php';
+        
+            $likes = new LikeArticle($idArticle, $idUser);
+            $totalUserLike = $likes->totalUserLike();
+        
+            if($totalUserLike == 0) {
+                $insertLike = new LikeArticle($idArticle, $idUser);
+    
+                if($insertLike->store()) {
+                    if($currentPage === $redirectPage) {
+                        header('location: blog.php');
+                    } else {
+                        header("location: detailsArticle.php?idArticle=$idArticle");
+                    }
+                }
+            } else {
+                $deleteLike = new LikeArticle($idArticle, $idUser);
+    
+                if($deleteLike->destroy()) {
+                    if($currentPage === $redirectPage) {
+                        header('location: blog.php');
+                    } else {
+                        header("location: detailsArticle.php?idArticle=$idArticle");
+                    }
+                }
+            } 
+        }
+
     }
 
 ?>
